Splitting admin functionallity modulary
